refactor: renamed weather and crypto client file
